combine duplicate config files
